moving on, slowing. I think all major ideas are setup
