+++ src/app.js
@@ -1,10 +1,17 @@
 import { appConfig } from './data/config.js'
 
-/** @type {any} */
+import { setupChallengesServer } from './content/challenges-server.js'
+import decodeMe from './content/decode-me.cjs'
+import mortalCoil from './content/mortal-coil.cjs'
+import survey from './server/routes/survey.cjs'
+import { htw } from './server/routes/htw.js'
+
+/** @type {any} App will be assembled step-wise*/
 const preApp = {
   config: appConfig,
 }
 
+// load parts of app
 require('./server/lib/withEntry')(preApp)
 require('./server/lib/withLogger')(preApp)
 require('./server/lib/withDb')(preApp)
@@ -17,6 +24,10 @@ require('./server/lib/withChallenges')(preApp)
 require('./server/lib/withStorage')(preApp)
 require('./server/lib/withChallengeStats')(preApp)
 
+/** @type {import('./data/types.js').App} */
+const App = preApp
+
+// load common functionality
 require('./server/lib/dbModel')(preApp)
 require('./server/lib/expressHeaders')(preApp)
 require('./server/lib/expressSession')(preApp)
@@ -31,12 +42,15 @@ require('./server/routes/user')(preApp)
 require('./server/routes/challenge')(preApp)
 require('./server/routes/setConfig')(preApp)
 
-/** @type {import('./data/types.js').App} */
-const App = preApp
+// htw routes/modules
+setupChallengesServer(App)
+decodeMe(App)
+mortalCoil(App)
+survey(App)
+htw(App)
 
 App.entry.start().then(() => {
   App.logger.info(App.moment().locale('en').format('LLLL'))
-  if (App.config.callback) App.config.callback(App)
 })
 
 if (process.env.UPTEST) {
