+++ src/data/config.js
@@ -1,11 +1,5 @@
-import { Sequelize } from 'sequelize'
 import { getI18nExtension } from '../content/i18n-extension.js'
 import { secrets } from '../helper/secrets-loader.js'
-import { setupChallengesServer } from '../content/challenges-server.js'
-
-const decodeMe = require('../content/decode-me.js')
-const mortalCoil = require('../content/mortal-coil.js')
-const survey = require('../server/routes/survey.js')
 
 export const appConfig = {
   database: (() => {
@@ -104,7 +98,7 @@ export const appConfig = {
     debug: false,
     fallbackLng: 'en',
     backend: {
-      loadPath: __dirname + '/lang/{{lng}}.json',
+      loadPath: process.cwd() + 'src/server/lang/{{lng}}.json',
     },
   },
   i18nExtend: getI18nExtension(),
@@ -123,282 +117,6 @@ export const appConfig = {
   editors: ['editor', 'demo'],
   noSelfAdmin: ['demo'],
   customCSS: '',
-  callback: function (/** @type {import('./types.js').App} */ App) {
-    setupChallengesServer(App)
-    decodeMe(App)
-    mortalCoil(App)
-    survey(App)
-
-    App.express.get('/news', (req, res) => {
-      res.renderPage({
-        page: 'news',
-      })
-    })
-
-    App.express.get('/hints', (req, res) => {
-      res.renderPage({
-        page: 'hints',
-      })
-    })
-
-    App.express.get('/links', (req, res) => {
-      res.renderPage({
-        page: 'links',
-      })
-    })
-
-    App.express.get('/api/top100', async (req, res) => {
-      const users = await App.db.models.User.findAll({
-        attributes: ['name', 'score', 'updatedAt'],
-        where: {
-          score: { [Sequelize.Op.gt]: 0 },
-        },
-        order: [
-          ['score', 'DESC'],
-          ['updatedAt', 'DESC'],
-        ],
-        limit: 100,
-        raw: true,
-      })
-      users.forEach((user, i) => {
-        if (i > 0 && users[i - 1].score == user.score) {
-          user.rank = users[i - 1].rank
-        } else {
-          user.rank = i + 1
-        }
-      })
-      res.json(users)
-    })
-
-    App.express.post('/api/user-rankings', async (req, res) => {
-      if (!req.body || !req.body.name || !Array.isArray(req.body.name)) {
-        res.send('bad body')
-        return
-      }
-      const names = req.body.name
-
-      try {
-        const users = await App.db.models.User.findAll({
-          attributes: [
-            'name',
-            'score',
-            'updatedAt',
-            [
-              Sequelize.literal(
-                '(SELECT COUNT(*) FROM Users as U WHERE U.score > User.score)'
-              ),
-              'rank',
-            ],
-          ],
-          where: {
-            name: names,
-          },
-          raw: true,
-        })
-        users.forEach((u) => {
-          u.rank += 1
-        })
-        res.json(users)
-      } catch (e) {
-        console.log(e)
-        res.send('db query failed')
-      }
-    })
-
-    App.express.get('/api/map', async (req, res) => {
-      res.json(
-        Object.keys(App.challenges.distance).filter(
-          (x) =>
-            x != secrets('secret_chal_1_id') && x != secrets('secret_chal_2_id')
-        )
-      )
-    })
-
-    /*App.express.get('/experiment', async (req, res) => {
-      const currentSolutions = await App.db.models.Solution.findAll({
-        where: {
-          updatedAt: {
-            [Sequelize.Op.gte]: App.moment().subtract(29, 'days').toDate(),
-          },
-        },
-        order: [['updatedAt']],
-        raw: true,
-      })
-
-      const persistenceScore = currentSolutions.reduce((result, obj) => {
-        const key = obj.UserId
-        const entry = (result[key] = result[key] ?? { mins: 0, lastSolved: -1 })
-        const ts = new Date(obj.updatedAt).getTime()
-        if (entry.lastSolved > 0) {
-          const diff = ts - entry.lastSolved
-          entry.mins += Math.min(30, Math.round(diff / 1000 / 60))
-        }
-        entry.lastSolved = ts
-        return result
-      }, {})
-
-      const persistenceArr = Object.entries(persistenceScore)
-
-      persistenceArr.sort((a, b) => b[1].mins - a[1].mins)
-
-      const top10 = persistenceArr.slice(0, 10)
-
-      for (const entry of top10) {
-        const user = await App.db.models.User.findOne({
-          where: { id: entry[0] },
-        })
-        console.log(user)
-        entry.name = user.name
-      }
-
-      console.log(top10)
-
-      res.send('hi')
-    })*/
-
-    /*App.express.get('/internal/newusers', async (req, res) => {
-      const usersFromDB = await App.db.models.User.findAll({
-        limit: 2000,
-        order: [['createdAt', 'DESC']],
-      })
-
-      function renderName(user) {
-        let output = escapeHtml(user.name)
-        if (user.score > 0) {
-          output = `<b>${output}</b>`
-        }
-        return output
-      }
-
-      const userStrings = usersFromDB.map(
-        (user) =>
-          `${App.moment(user.createdAt).format(
-            'DD.MM.YYYY HH:mm'
-          )}: ${renderName(user)}, ${user.score} Punkte`
-      )
-      res.send(userStrings.join('<br>'))
-    })*/
-
-    if (process.env.SAVE2LOCAL && !process.env.UBERSPACE) {
-      run()
-
-      async function run() {
-        if (!process.env.LIVE) throw 'NOT CONNECTED TO LIVE SERVER'
-        const LOCALAPP = {}
-
-        // @ts-ignore
-        LOCALAPP.db = new Sequelize({
-          dialect: 'sqlite',
-          storage: './db.sqlite',
-          logging: false,
-        })
-        await require('../server/lib/dbModel.js')(LOCALAPP)
-        await LOCALAPP.db.authenticate()
-
-        // Es ist viel schneller, die gesamte Datenbank neu aufzusetzen
-        await LOCALAPP.db.sync({ force: true })
-
-        console.log('Lokale Datenbank synchronisiert')
-
-        console.log('Starte Import Räume ...')
-
-        const rooms = await App.db.models.Room.findAll({ raw: true })
-        await LOCALAPP.db.models.Room.bulkCreate(rooms, { silent: true })
-
-        console.log('Starte Import Benutzer und gelöste Aufgaben ...')
-
-        console.log('  Lade Nutzer von Server')
-        const users = await App.db.models.User.findAll({ raw: true })
-
-        console.log('  Lade Lösungen vom Server')
-        const solutions = await App.db.models.Solution.findAll({ raw: true })
-
-        console.log(`  Füge ${users.length} Nutzer lokal ein`)
-        await LOCALAPP.db.models.User.bulkCreate(users)
-
-        console.log(`  Füge ${solutions.length} Lösungen lokal ein`)
-        await LOCALAPP.db.models.Solution.bulkCreate(solutions)
-
-        console.log('Starte Import KVPairs ...')
-
-        const kvpairs = await App.db.models.KVPair.findAll({ raw: true })
-
-        // sqlite not supporting null bytes in strings
-        for (const pair of kvpairs) {
-          pair.value = pair.value.replace(/\0/g, '')
-        }
-
-        await LOCALAPP.db.models.KVPair.bulkCreate(kvpairs)
-
-        console.log('  KVPAirs vollständig')
-
-        process.exit()
-      }
-    }
-
-    if (!process.env.UBERSPACE) {
-      require('../server/routes/analyze.js')(App)
-    }
-
-    if (process.env.RECALCULATESCORE) {
-      void (async () => {
-        console.log('\nStart recalculating scores')
-        const users = await App.db.models.User.findAll()
-
-        const solutions = await App.db.models.Solution.findAll({ raw: true })
-
-        // make sure data is consistent by retrieving scores again and compare
-        const users2 = await App.db.models.User.findAll({ raw: true })
-        const userScores1 = {}
-        for (const user of users) {
-          userScores1[user.id] = user.score
-        }
-        for (const user of users2) {
-          if (user.score !== userScores1[user.id]) {
-            console.log(
-              `user ${user.name} solved a challenge while retrieving data, making data inconsistent. Please rerun.`
-            )
-            process.exit(1)
-          }
-        }
-
-        const byUser = {}
-
-        solutions.forEach((sol) => {
-          if (!byUser[sol.UserId]) byUser[sol.UserId] = []
-
-          byUser[sol.UserId].push(sol.cid)
-        })
-
-        let hasChange = false
-
-        for (const user of users) {
-          const solutions = byUser[user.id] ?? []
-          let score = 0
-          for (const solution of solutions) {
-            if (App.challenges.data.some((c) => c.id == solution)) {
-              score += 10 + (App.challenges.distance[solution] || 0)
-            }
-          }
-          if (user.score != score) {
-            hasChange = true
-            console.log(`${user.score} -> ${score}`)
-          }
-          user.score = score
-          await user.save({ silent: true })
-        }
-        console.log('completed')
-
-        if (hasChange) {
-          console.log(
-            'changes saved. to make sure data is consistent, please rerun'
-          )
-          process.exit(1)
-        }
-        process.exit()
-      })()
-    }
-  },
   masterPassword: secrets('config_master_password'),
   githubHref: '/links',
   githubTargetBlank: false,
