+++ tsconfig.json
@@ -3,8 +3,14 @@
   "compilerOptions": {
     "checkJs": true,
     "noEmit": true,
-    "noImplicitAny": false,
-    "esModuleInterop": true
+    "noImplicitAny": false
   },
-  "include": ["./src/**/*.js", "./src/**/*.ts", "src/helper/secrets-loader.js"]
+  "include": [
+    "./src/**/*.js",
+    "./src/**/*.ts",
+    "src/helper/secrets-loader.js",
+    "src/content/decode-me.cjs",
+    "src/content/mortal-coil.cjs",
+    "src/server/routes/survey.cjs"
+  ]
 }
